reorganize util files
