+++ fortran/test/unit/util.F90
@@ -4,7 +4,7 @@
 ! Tests for the musica_util module
 program test_util
 
-#include "musica/error.hpp"
+#include "musica/utils/error.hpp"
 
 #define ASSERT( expr ) call assert( expr, __FILE__, __LINE__ )
 #define ASSERT_EQ( a, b ) call assert( a == b, __FILE__, __LINE__ )
