+++ include/musica/utils/error.hpp
