+++ src/CMakeLists.txt
@@ -26,7 +26,9 @@ add_library(musica::musica ALIAS musica)
 
 # Common sources
 target_sources(musica PRIVATE
-  ${CMAKE_CURRENT_SOURCE_DIR}/util.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/utils/error_handler.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/utils/string.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/utils/util.cpp
   ${CMAKE_BINARY_DIR}/version.cpp
 )
